// elza/elza-react/src/api/generated/model/ae-part-form-vo.ts
// tslint:disable
/**
 * CLIENT API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { AeItemVO } from './ae-item-vo';
import { PartType } from './part-type';

/**
 * Formulář pro odeslání hodnot atributů popisující část archivní entity
 * @export
 * @interface AePartFormVO
 */
export interface AePartFormVO {
    /**
     *
     * @type {PartType}
     * @memberof AePartFormVO
     */
    part: PartType;
    /**
     * Seznam všech hodnot atributů (existující, nové)
     * @type {Array<AeItemVO>}
     * @memberof AePartFormVO
     */
    items: Array<AeItemVO>;
    /**
     * Identifikátor nadřízené části
     * @type {number}
     * @memberof AePartFormVO
     */
    parentPartId?: number;
    /**
     * Identifikátor části
     * @type {number}
     * @memberof AePartFormVO
     */
    partId?: number;
}