// elza/elza-react/src/api/generated/model/ae-part-vo.ts
// tslint:disable
/**
 * CLIENT API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { AeItemVO } from './ae-item-vo';

/**
 * Předek pro hodnotu partu
 * @export
 * @interface AePartVO
 */
export interface AePartVO {
    /**
     * Discriminator
     * @type {string}
     * @memberof AePartVO
     */
    '@class': string;
    /**
     * Interní identifikátor
     * @type {number}
     * @memberof AePartVO
     */
    id?: number;
    /**
     * Jednoznačný identifikátor hodnoty atributu
     * @type {string}
     * @memberof AePartVO
     */
    uuid?: string;
    /**
     * Textová reprezentace partu
     * @type {string}
     * @memberof AePartVO
     */
    textValue: string;
    /**
     * Identifikátor nadřazené části
     * @type {number}
     * @memberof AePartVO
     */
    parentPartId?: number;
    /**
     * Jednoznačný identifikátor nadřazené části
     * @type {string}
     * @memberof AePartVO
     */
    parentPartUuid?: string;
    /**
     * Strukturované hodnoty atributů
     * @type {Array<AeItemVO>}
     * @memberof AePartVO
     */
    items: Array<AeItemVO>;
}